+++ packages/biliass/rust/src/convert.rs
@@ -69,8 +69,8 @@ pub fn convert_to_ass<Reader, Input>(
     text_opacity: f32,
     duration_marquee: f64,
     duration_still: f64,
-    block_options: &BlockOptions,
     is_reduce_comments: bool,
+    block_options: &BlockOptions,
 ) -> Result<String, BiliassError>
 where
     Reader: Fn(Input, f32, (f32, f32, f32), &BlockOptions) -> Result<Vec<Comment>, BiliassError>
@@ -82,10 +82,9 @@ where
         crate::reader::special::BILI_PLAYER_SIZE,
         (stage_width, stage_height),
     );
-    // let block_options = BlockOptions::default();
     let comments_result: Result<Vec<Vec<Comment>>, BiliassError> = inputs
         .into_par_iter()
-        .map(|input| reader(input, font_size, zoom_factor, &block_options))
+        .map(|input| reader(input, font_size, zoom_factor, block_options))
         .collect();
 
     let comments = comments_result?;
