+++ packages/biliass/rust/src/python/convert.rs
@@ -1,14 +1,80 @@
 use crate::error::BiliassError;
 use crate::filter::BlockOptions;
-use crate::python::PyBlockOptions;
 use crate::{convert, reader};
 use pyo3::{
-    conversion,
     prelude::*,
     pybacked::{PyBackedBytes, PyBackedStr},
-    types::PyDict,
 };
 use regex::Regex;
+
+#[pyclass(name = "BlockOptions")]
+#[derive(Clone)]
+pub struct PyBlockOptions {
+    pub block_top: bool,
+    pub block_bottom: bool,
+    pub block_scroll: bool,
+    pub block_reverse: bool,
+    pub block_special: bool,
+    pub block_colorful: bool,
+    pub block_keyword_patterns: Vec<String>,
+}
+
+impl PyBlockOptions {
+    pub fn to_block_options(&self) -> Result<BlockOptions, BiliassError> {
+        let block_keyword_patterns_res: Result<Vec<Regex>, regex::Error> = self
+            .block_keyword_patterns
+            .iter()
+            .map(|pattern| regex::Regex::new(pattern))
+            .collect();
+        let block_keyword_patterns = block_keyword_patterns_res.map_err(BiliassError::from)?;
+        Ok(BlockOptions {
+            block_top: self.block_top,
+            block_bottom: self.block_bottom,
+            block_scroll: self.block_scroll,
+            block_reverse: self.block_reverse,
+            block_special: self.block_special,
+            block_colorful: self.block_colorful,
+            block_keyword_patterns,
+        })
+    }
+}
+
+#[pymethods]
+impl PyBlockOptions {
+    #[new]
+    fn new(
+        block_top: bool,
+        block_bottom: bool,
+        block_scroll: bool,
+        block_reverse: bool,
+        block_special: bool,
+        block_colorful: bool,
+        block_keyword_patterns: Vec<String>,
+    ) -> PyResult<Self> {
+        Ok(PyBlockOptions {
+            block_top,
+            block_bottom,
+            block_scroll,
+            block_reverse,
+            block_special,
+            block_colorful,
+            block_keyword_patterns,
+        })
+    }
+
+    #[staticmethod]
+    pub fn default() -> Self {
+        PyBlockOptions {
+            block_top: false,
+            block_bottom: false,
+            block_scroll: false,
+            block_reverse: false,
+            block_special: false,
+            block_colorful: false,
+            block_keyword_patterns: vec![],
+        }
+    }
+}
 
 #[pyclass(name = "ConversionOptions")]
 pub struct PyConversionOptions {
@@ -25,6 +91,7 @@ pub struct PyConversionOptions {
 }
 
 #[pymethods]
+#[allow(clippy::too_many_arguments)]
 impl PyConversionOptions {
     #[new]
     fn new(
@@ -50,12 +117,10 @@ impl PyConversionOptions {
             duration_still,
             is_reduce_comments,
             block_options: block_options.clone(),
-            // block_options: PyBlockOptions::default(),
         }
     }
 }
 
-#[allow(clippy::too_many_arguments)]
 #[pyfunction(name = "xml_to_ass")]
 pub fn py_xml_to_ass(
     inputs: Vec<PyBackedStr>,
@@ -72,12 +137,11 @@ pub fn py_xml_to_ass(
         conversion_options.text_opacity,
         conversion_options.duration_marquee,
         conversion_options.duration_still,
-        &conversion_options.block_options.to_block_options()?,
         conversion_options.is_reduce_comments,
+        &conversion_options.block_options.to_block_options()?,
     )?)
 }
 
-#[allow(clippy::too_many_arguments)]
 #[pyfunction(name = "protobuf_to_ass")]
 pub fn py_protobuf_to_ass(
     inputs: Vec<PyBackedBytes>,
@@ -94,7 +158,7 @@ pub fn py_protobuf_to_ass(
         conversion_options.text_opacity,
         conversion_options.duration_marquee,
         conversion_options.duration_still,
-        &conversion_options.block_options.to_block_options()?,
         conversion_options.is_reduce_comments,
+        &conversion_options.block_options.to_block_options()?,
     )?)
 }
