+++ packages/biliass/rust/src/python/mod.rs
@@ -1,8 +1,6 @@
 mod convert;
-mod filter;
 mod logging;
 mod proto;
-pub use convert::{py_protobuf_to_ass, py_xml_to_ass, PyConversionOptions};
-pub use filter::PyBlockOptions;
+pub use convert::{py_protobuf_to_ass, py_xml_to_ass, PyBlockOptions, PyConversionOptions};
 pub use logging::py_enable_tracing;
 pub use proto::py_get_danmaku_meta_size;
